communication/entities/python/collector/requirements.txt
paho-mqtt>=1.2
MySQL-python>1.2.4
protobuf>=3.1.0
